Update to use VS2015.
GL
- GLFramebufferObject now consistently uses a drawbuffer index.
- GLSeparableProgram and GLProgramPipeline are no longer copyable.

apps/rgbd_calibration
- Updated to use gflags.

apps/picture_taker
- Updated to use gflags.
- DepthAverager now outputs a PFM average that can be loaded as a point
  cloud.

+++ core/src/io/OBJLoader.cpp
@@ -597,7 +597,7 @@ bool OBJLoader::getVertexAttributes( const std::string& objFaceVertexToken,
 
     std::vector< std::string > vertexAttributes;
     pystring::split( objFaceVertexToken, vertexAttributes, "/" );
-    int numVertexAttributes = vertexAttributes.size();
+    size_t numVertexAttributes = vertexAttributes.size();
 
     // Check if it has position. It is required.
     if( numVertexAttributes < 1 )
